+++ js/index.js
@@ -1,5 +1,6 @@
 export * from "./base";
 export * from "./repair";
+export * from "./merchandise";
 export * from "./repair-reference";
 export * from "./sale-snapshot";
 export * from "./store";